+++ paper/paper.bib
@@ -7,7 +7,8 @@
   year      = {2005},
   editor    = {Charles D. Hansen and Christopher R. Johnson},
   pages     = {717--731},
-  isbn      = {978-0123875822}
+  isbn      = {978-0123875822},
+  doi       = {10.1016/B978-012387582-2/50038-1}
 }
 
 @InProceedings{vapor,
@@ -122,6 +123,7 @@ address={San Rafael, CA}
     volume = {30},
     pages = {082102},
     year = {2023},
+    doi = {10.1063/5.0148156}
 }
 
 @Software{Brandenburg-2020-Zenodo,
@@ -150,7 +152,8 @@ address={San Rafael, CA}
   journal={Applied Physics Letters},
   volume={123},
   year={2023},
-  publisher={AIP Publishing}
+  publisher={AIP Publishing},
+  doi={10.1063/5.0152083}
 }
 
 @software{candelaresi_2025_15988189,
@@ -172,7 +175,7 @@ volume = {130},
 pages = {104264},
 year = {2025},
 issn = {0097-8493},
-doi = {https://doi.org/10.1016/j.cag.2025.104264},
+doi = {10.1016/j.cag.2025.104264},
 url = {https://www.sciencedirect.com/science/article/pii/S0097849325001050},
 author = {José Marín and Tiffany M.G. Baptiste and Cristobal Rodero and Steven E. Williams and Steven A. Niederer and Ignacio García-Fernández},
 }
