### 2026-07-15 f09cd47
Split "Statement of need" into "Statement of need" and "State of the field". No new text there, just split after the second paragraph.
Add a Research impact statement.

Add AI usage disclosure.

Add/fix some DOI numbers.
