+++ paper/paper.bib
@@ -97,7 +97,7 @@ address={San Rafael, CA}
 @article{kedia2013tying,
   title={Tying knots in light fields},
   author={Kedia, Hridesh and Bialynicki-Birula, Iwo and Peralta-Salas, Daniel and Irvine, William TM},
-  journal={Phys.\ Rev.\ Lett.},
+  journal={Phys. Rev. Lett.},
   volume={111},
   pages={150404},
   year={2013},
@@ -108,7 +108,7 @@ address={San Rafael, CA}
 @article{smiet2019mapping,
   title={Mapping the sawtooth},
   author={Smiet, CB and Kramer, GJ and Hudson, SR},
-  journal={Plasma Phys.\ Contr.\ F.},
+  journal={Plasma Phys. Contr. F.},
   volume={62},
   pages={025007},
   year={2019},
@@ -170,7 +170,7 @@ address={San Rafael, CA}
 
 @article{SciBlend,
 title = {SciBlend: Advanced data visualization workflows within Blender},
-journal = {Comput.\ Graphics},
+journal = {Comput. Graphics},
 volume = {130},
 pages = {104264},
 year = {2025},
@@ -178,4 +178,15 @@ issn = {0097-8493},
 doi = {10.1016/j.cag.2025.104264},
 url = {https://www.sciencedirect.com/science/article/pii/S0097849325001050},
 author = {José Marín and Tiffany M.G. Baptiste and Cristobal Rodero and Steven E. Williams and Steven A. Niederer and Ignacio García-Fernández},
+}
+
+@article{vortex_braids,
+    author = {Candelaresi, S. and Hornig, G. and Podger, B. and Pontin, D. I.},
+    title = {Topological constraints in the reconnection of vortex braids},
+    journal = {Phys. Fluids},
+    volume = {33},
+    number = {5},
+    pages = {056101},
+    year = {2021},
+    doi = {10.1063/5.0047033},
 }
