### 2026-08-21 #11181 @editorialbot
[REVIEW]: BlenDaViz, scientific visualization library for Blender in Python
**Submitting author:** @SimonCan (<a href="http://orcid.org/0000-0002-7666-8504">Simon Candelaresi</a>)
**Repository:** https://github.com/SimonCan/BlenDaViz
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.0
