### 2026-08-21 #11181 @jolars
:wave: @Miko997, @Chilipp, @MaaniBeigy – Thanks for agreeing to review this submission. This is the review thread for the paper. All of our communications will happen here from now on.

As you can see above, you each should use the command

```@editorialbot generate my checklist```

to create your review checklist. @editorialbot commands need to be the first thing in a new comment.

We aim for reviews to be completed within about 3 weeks. Please let me know if any of you require some more time. We can also use EditorialBot (our bot) to set automatic reminders if you know you'll be away for a known period of time.
