### 2026-08-30 SimonCan/BlenDaViz#1 @Miko997
Official Blender Snap installation is not importable without --python-use-system-env
While reviewing BlenDaViz for JOSS, I ran into an installation-path issue, at least on my setup:

Ubuntu 24.04.3, x86_64
Blender 5.2.1 LTS from the official Snap
Commit f09cd47c535f9720aaac94791bceb3979f01967e

The documented Blender-Python pip install . command completed successfully, but pip installed the package into my user site-packages directory. A normal Blender launch did not include that directory, so import blendaviz failed with ModuleNotFoundError.

Launching Blender with:

blender --python-use-system-env

made the import work. The quick-start example rendered, and all six automated tests passed with the same option. Without it, the test runner also failed to import coverage after installing it into the user site.

Could the installation documentation mention this for Snap installations, or provide a method that works with a normal Blender launch?

Related JOSS review: openjournals/joss-reviews#11181
