### 2026-08-30 #11181 @Miko997
While reviewing BlenDaViz, I found an installation and test-environment issue on my Ubuntu 24.04.3 setup using the official Blender 5.2.1 Snap. The documented installation completed, but a normal Blender launch could not import the user-site installation. With --python-use-system-env, the import worked and all six tests passed. Details: SimonCan/BlenDaViz#1
