### 2026-08-30 SimonCan/BlenDaViz#1 @Miko997
For completeness, these were the exact results on my setup:

- Normal Blender import: normal_blender_exit=7
- Import with --python-use-system-env: system_env_import_exit=0
- Test runner without the option: normal_test_exit=7, failing to import coverage
- Test runner with the option: workaround_test_exit=0
- Six tests ran successfully in 21.176 seconds with the workaround

I have preserved the complete command output locally if further details are needed.
