### 2026-08-30 SimonCan/BlenDaViz#2 @Miko997
Test runner does not propagate unsuccessful test results to CI
While reviewing BlenDaViz for JOSS, I noticed that failed unittest results do not currently make the CI job fail.

In one CI run, Blender 4.3.2 reported `FAILED (errors=1)`, but the GitHub Actions job still completed successfully. I understand that Blender 4.3.2 is below the supported minimum. the issue is only that the test result is not propagated.

Could `run_blender_tests.py` exit with a non-zero status when `result.wasSuccessful()` is false?

Related JOSS review: openjournals/joss-reviews#11181
