### 2026-09-02 a9208c9
Make CI job fail if unit test fails.
