### 2026-09-02 SimonCan/BlenDaViz#2 @SimonCan
Thanks for spotting this. I fixed this in this commit: [https://github.com/SimonCan/BlenDaViz/commit/a9208c986fd6d18d8eeae75b452533befd4ed022](https://github.com/SimonCan/BlenDaViz/commit/a9208c986fd6d18d8eeae75b452533befd4ed022).
