### 2026-09-02 fdd9a94
Add a note about snap installed Blender and the usage of the additional parameter:
blender --python-use-system-env
