### 2026-09-02 SimonCan/BlenDaViz#1 @SimonCan
Fixed this here: [https://github.com/SimonCan/BlenDaViz/commit/fdd9a943061af0f3e6f87ca21424d1f555624e6c](https://github.com/SimonCan/BlenDaViz/commit/fdd9a943061af0f3e6f87ca21424d1f555624e6cl).
