### 2026-09-03 cd33b2f
Fix error from shape (1,) return from RegularGridInterpolator.
