### 2026-09-05 SimonCan/BlenDaViz#2 @Miko997
Thanks, retested at cd33b2f on Blender 5.2.1 LTS using --python-use-system-env --python-exit-code 1. All six original tests passed with exit code 0. In a separate copy, adding one deliberately failing test produced seven tests with one failure and exit code 1. This resolves the exit-code issue on my setup.
