### 2026-09-05 SimonCan/BlenDaViz#1 @Miko997
Thanks, retested at cd33b2f with the official Blender 5.2.1 Snap. The Snap instructions are now included in the README. Installation completed successfully, and import worked with the documented --python-use-system-env option. This resolves the issue on my setup.
