### 2026-09-05 SimonCan/BlenDaViz#3 @Miko997
Contour plot fails with the default color value
While reviewing BlenDaViz for JOSS at cd33b2f, I ran into an issue with the documented/default contour call on Blender 5.2.1.

c = blt.contour(phi, xx, yy, zz) fails with:

AttributeError: 'tuple' object has no attribute 'ndim'

Passing the same color explicitly as a NumPy array works and renders successfully:

c = blt.contour(phi, xx, yy, zz, color=np.array([0, 1, 0, 1]))

It looks like the default tuple color reaches code that expects .ndim.

Related JOSS review: openjournals/joss-reviews#11181
