### 2026-09-05 #11181 @Miko997
During the manual functionality review I found one additional issue with the documented/default contour call on Blender 5.2.1. Details: SimonCan/BlenDaViz#3
