### 2026-09-05 SimonCan/BlenDaViz#4 @Miko997
Matplotlib embedding renders incorrectly without pixel normalization
While reviewing BlenDaViz for JOSS at cd33b2f on Blender 5.2.1, I noticed that mpl_figure_to_blender() completes successfully, but the embedded Matplotlib figure is almost washed out in the Blender render.

I reproduced this again in a fresh Blender session. The source Matplotlib image looked correct, while the BlenDaViz render did not. Normalizing the image values from 0–255 to 0–1 before assigning them to the Blender texture made the render look correct.

Related JOSS review: openjournals/joss-reviews#11181
