### 2026-09-05 #11181 @Miko997
During the manual functionality review I found an additional issue with Matplotlib embedding on Blender 5.2.1. Details: SimonCan/BlenDaViz#4
