### 2026-09-05 SimonCan/BlenDaViz#5 @Miko997
JOSS review: remaining documentation and time-update observations
Thanks for the installation and test-runner fixes. I’ve finished the planned runtime checks at `cd33b2f` on Blender 5.2.1. Apart from the findings already tracked in #3 and #4, I have a few remaining notes.

**Time updates**
In my two-frame `PathLine` test, changing from frame 1 to 2 left the plot unchanged. Calling `pl.plot()` manually then updated it correctly, moving the first point from x=0 to x=10. I found no `frame_change_pre` handler registered for that object. Could you check this behavior or clarify in the paper which plot types update automatically?

**Documentation**
The streamline example calls `blt.streamlines(...)`, while the available functions are `streamlines_function` and `streamlines_array`. The script command described as running without the GUI also omits `--background`. Could these examples be updated?

**Small paper suggestions**
Using the printed PDF line numbers:
- Line 66: “Every of our plotting functions” → “Each of our plotting functions”.
- Line 69: “a texture is add” → “a texture is added”.
- Line 117: the Matplotlib reference renders as “& al., et.”; please check the author formatting.

I’m collecting these here rather than opening a separate issue for each point.

Related JOSS review: openjournals/joss-reviews#11181
