### 2026-09-05 #11181 @Miko997
My planned runtime checks at `cd33b2f` on Blender 5.2.1 are complete. The remaining reported findings are tracked in SimonCan/BlenDaViz#3, SimonCan/BlenDaViz#4 and SimonCan/BlenDaViz#5.

I’ll verify the relevant changes when they’re ready. My overall review is still in progress.
