### 2026-09-07 1535c77
Color tuples are noew converted into Numpy arrays to avoid error when checking .ndim.
