### 2026-09-07 226b9a7
Register time_handler for 1d line/marker plots for time dependent data.
