### 2026-09-07 3fdeca0
Fix the streamlines_array call signature in the documentation, as well as the background plot.
