### 2026-09-07 30c860e
Fix typos in paper.
