+++ paper/paper.bib
@@ -49,7 +49,7 @@ DOI="10.1007/978-3-642-03915-7"
 title={matplotlib/matplotlib: REL: v3.2.1},
 DOI={10.5281/zenodo.3714460},
 publisher={Zenodo},
-author={Thomas A Caswell and Michael Droettboom and Antony Lee and John Hunter and Eric Firing and David Stansby and Jody Klymak and Tim Hoffmann and Elliott Sales de Andrade and Nelle Varoquaux and et al.},
+author={Thomas A Caswell and Michael Droettboom and Antony Lee and John Hunter and Eric Firing and David Stansby and Jody Klymak and Tim Hoffmann and Elliott Sales de Andrade and Nelle Varoquaux and others},
 year={2020},
 month={Mar}
 }
