### 2026-09-07 e17a12a
Fix replot of marker in case previous geometry is not selected.
