### 2026-09-07 80187ee
Check if color_rgba has ndim == 1.
