### 2026-09-09 875bfc6
Correct DOI for Gruchalla et al. to point to the book chapter rather than the entire book.
