+++ paper/paper.bib
@@ -28,7 +28,7 @@ publisher="Springer Berlin Heidelberg",
 address="Berlin, Heidelberg",
 pages="321--332",
 isbn="978-3-642-03915-7",
-DOI="10.1007/978-3-642-03915-7"
+DOI="10.1007/978-3-642-03915-7_28"
 }
 
 @Article{Hunter2007,
